Fix inability to read files exceeding int32 size limit

+++ cpp/src/file/read_file.h
@@ -37,19 +37,19 @@ class ReadFile {
 
     int open(const std::string& file_path);
     FORCE_INLINE bool is_opened() const { return fd_ > 0; }
-    FORCE_INLINE int32_t file_size() const { return file_size_; }
+    FORCE_INLINE int64_t file_size() const { return file_size_; }
     FORCE_INLINE const std::string& file_path() const { return file_path_; }
 
     /*
      * try to reader @buf_size bytes from @offset of this file
      * into @buf. @read_len return the actual len reader.
      */
-    int read(int32_t offset, char* buf, int32_t buf_size,
+    int read(int64_t offset, char* buf, int32_t buf_size,
              int32_t& ret_read_len);
     void close();
 
    private:
-    int get_file_size(int32_t& file_size);
+    int get_file_size(int64_t& file_size);
     int check_file_magic();
 
    private:
@@ -59,7 +59,7 @@ class ReadFile {
    private:
     std::string file_path_;
     int fd_;
-    int32_t file_size_;
+    int64_t file_size_;
 };
 
 }  // end namespace storage
