+++ cpp/src/reader/chunk_reader.cc
@@ -228,7 +228,7 @@ int ChunkReader::read_from_file_and_rewrap(int want_size) {
     int ret = E_OK;
     const int DEFAULT_READ_SIZE = 4096;  // may use page_size + page_header_size
     char* file_data_buf = in_stream_.get_wrapped_buf();
-    int offset = chunk_meta_->offset_of_chunk_header_ + chunk_visit_offset_;
+    int64_t offset = chunk_meta_->offset_of_chunk_header_ + chunk_visit_offset_;
     int read_size =
         (want_size < DEFAULT_READ_SIZE ? DEFAULT_READ_SIZE : want_size);
     if (file_data_buf_size_ < read_size ||
